rename ToolCardsCarouselView
